requirements.txt
flask
openai==0.28
faiss-cpu
numpy
